+++ app/controllers/order_dispatches_controller.rb
@@ -348,8 +348,8 @@ class OrderDispatchesController < ApplicationController
     product = order_item.channel_product.product_mapping.product
     return multipack_unallocation(order_item, product) if product.product_type.eql? 'multiple'
 
-    product.update(available_stock: product.available_stock.to_f + order_item.ordered,
-                   allocated: product.allocated.to_f - order_item.ordered, allocated_orders: product.allocated_orders.to_i - 1)
+    product.update(available_stock: product.available_stock.to_i + order_item.ordered,
+                   allocated: product.allocated.to_i - order_item.ordered, allocated_orders: product.allocated_orders.to_i - 1)
     #  change_log: "#{order_item.channel_order.channel_type} API, #{order_item.channel_order.id}, #{order_item.channel_order.order_id}, UnAllocate, #{order_item.channel_product.listing_id}")
     order_item.update(allocated: false)
     flash[:notice] = 'Unallocation successful!'
@@ -360,8 +360,8 @@ class OrderDispatchesController < ApplicationController
     return multipack_allocation(order_item, product) if product&.product_type.eql? 'multiple'
 
     if product&.available_stock.to_i >= order_item.ordered
-      product.update(available_stock: product.available_stock.to_f - order_item.ordered,
-                     allocated: product.allocated.to_f + order_item.ordered, allocated_orders: product.allocated_orders.to_i + 1)
+      product.update(available_stock: product.available_stock.to_i - order_item.ordered,
+                     allocated: product.allocated.to_i + order_item.ordered, allocated_orders: product.allocated_orders.to_i + 1)
       # change_log: "#{order_item.channel_order.channel_type} API, #{order_item.channel_order.id}, #{order_item.channel_order.order_id}, Allocated, #{order_item.channel_product.listing_id}")
       order_item.update(allocated: true)
     else
@@ -374,8 +374,8 @@ class OrderDispatchesController < ApplicationController
       child = multipack.child
       quantity = multipack.quantity
       ordered = (order_item.ordered * quantity)
-      child.update(available_stock: child.available_stock.to_f + ordered,
-                   allocated: child.allocated.to_f - ordered, allocated_orders: child.allocated_orders.to_i - 1)
+      child.update(available_stock: child.available_stock.to_i + ordered,
+                   allocated: child.allocated.to_i - ordered, allocated_orders: child.allocated_orders.to_i - 1)
       # ,change_log: "#{order_item.channel_order.channel_type} API, #{order_item.channel_order.id}, #{order_item.channel_order.order_id}, UnAllocate, #{order_item.channel_product.listing_id}"
     end
     order_item.update(allocated: false)
@@ -391,8 +391,8 @@ class OrderDispatchesController < ApplicationController
         child = multipack.child
         quantity = multipack.quantity
         ordered = (order_item.ordered * quantity)
-        child.update(available_stock: child.available_stock.to_f - ordered,
-                     allocated: child.allocated.to_f + ordered, allocated_orders: child.allocated_orders.to_i + 1)
+        child.update(available_stock: child.available_stock.to_i - ordered,
+                     allocated: child.allocated.to_i + ordered, allocated_orders: child.allocated_orders.to_i + 1)
         # change_log: "#{order_item.channel_order.channel_type} API, #{order_item.channel_order.id}, #{order_item.channel_order.order_id}, Allocated, #{order_item.channel_product.listing_id}")
       end
       order_item.update(allocated: true)
@@ -402,8 +402,8 @@ class OrderDispatchesController < ApplicationController
   end
 
   def allocate(product, ordered)
-    product.update(available_stock: product.available_stock.to_f - ordered,
-                   allocated: product.allocated.to_f + ordered)
+    product.update(available_stock: product.available_stock.to_i - ordered,
+                   allocated: product.allocated.to_i + ordered)
     #  change_log: "#{order_item.channel_order.channel_type} API, #{order_item.channel_order.id}, #{order_item.channel_order.order_id}, Allocated, #{order_item.channel_product.listing_id}")
   end
 
@@ -509,15 +509,15 @@ class OrderDispatchesController < ApplicationController
         child.update(
           change_log: "Cancel Order, #{order_item.channel_product.item_sku}, #{order_item.channel_order.order_id},
           Cancel Order, #{order_item.channel_product.listing_id}, #{ordered}, #{inventory_balance},
-          #{channel_type}", available_stock: child.available_stock.to_f + ordered,
+          #{channel_type}", available_stock: child.available_stock.to_i + ordered,
           unshipped: child.unshipped.to_i - ordered, unshipped_orders: child.unshipped_orders.to_i - 1,
-          allocated: child.allocated.to_f - ordered, allocated_orders: child.allocated_orders.to_i - 1
+          allocated: child.allocated.to_i - ordered, allocated_orders: child.allocated_orders.to_i - 1
         )
       else
         child.update(
           change_log: "Cancel Order, #{order_item.channel_product.item_sku}, #{order_item.channel_order.order_id},
           Cancel Order, #{order_item.channel_product.listing_id}, #{ordered}, #{inventory_balance},
-          #{channel_type}", available_stock: child.available_stock.to_f + ordered,
+          #{channel_type}", available_stock: child.available_stock.to_i + ordered,
           unshipped: child.unshipped.to_i - ordered, unshipped_orders: child.unshipped_orders.to_i - 1
         )
       end
