+++ app/controllers/trackings_controller.rb
@@ -190,7 +190,7 @@ class TrackingsController < ApplicationController
           product = multi.child
           # quantity = multi.quantity.to_f * (product.pack_quantity.nil? ? 1 : product.pack_quantity.to_f)
           # products << { sku: product.sku, product: product, quantity: quantity * multiple_product.ordered }
-          products << { sku: product.sku, product: product, quantity: multi.quantity.to_f * multiple_product.ordered }
+          products << { sku: product.sku, product: product, quantity: multi.quantity.to_i * multiple_product.ordered }
         end
       end
