+++ app/jobs/issue_orders_job.rb
@@ -36,7 +36,7 @@ class IssueOrdersJob < ApplicationJob
       if product.present?
         next multipack_product(item, product) unless product.product_type.eql? 'single'
 
-        inventory_balance = product.inventory_balance.to_f - item.ordered
+        inventory_balance = product.inventory_balance.to_i - item.ordered
         update_available_stock(item, product, inventory_balance, item.ordered)
       end
     end
@@ -56,7 +56,7 @@ class IssueOrdersJob < ApplicationJob
       quantity = multipack.quantity
       child = multipack.child
 
-      inventory_balance = child.inventory_balance.to_f - (item.ordered * quantity)
+      inventory_balance = child.inventory_balance.to_i - (item.ordered * quantity)
       update_available_stock(item, child, inventory_balance, (item.ordered * quantity))
     end
   end
